+++ src/interactions/collection.js
@@ -1,4 +1,3 @@
-import Interactor from '../interactor';
 import interactor from '../decorator';
 
 /**
@@ -27,6 +26,18 @@ import interactor from '../decorator';
  * await checkboxGroup
  *   .item(0).click()
  *   .item(1).click()
+ * ```
+ *
+ * Nested interactors also have an additional method, `#only()`, which
+ * disables the default nested chaining behavior, but retains any
+ * previous interactions.
+ *
+ * ``` javascript
+ * await checkboxGroup
+ *   .item(0).click()
+ *   .item(1).only()
+ *     .focus()
+ *     .trigger('keydown', { which: 32 })
  * ```
  *
  * When calling a collection method without an index, an array of
@@ -89,18 +100,7 @@ import interactor from '../decorator';
  * @returns {Object} Property descriptor
  */
 export default function(selector, descriptors = {}) {
-  let ItemInteractor;
-
-  // if an interactor was provided, use it
-  if (descriptors.prototype instanceof Interactor) {
-    ItemInteractor = descriptors;
-
-    // otherwise, create a new one
-  } else {
-    ItemInteractor = interactor(function() {
-      Object.assign(this, descriptors);
-    });
-  }
+  let ItemInteractor = interactor(descriptors);
 
   return function(index) {
     // when no index is provided, map all elements to interactors