+++ CHANGELOG.md
@@ -5,10 +5,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+- `scoped` property creator for scoped interactors
+- `only` method so nested interactors can break out of parent chains
+
 ### Changed
 
 - relax restrictions around reserved interactor properties
 - default properties can be freely overwritten
+- interactor decorator to support pojos
 
 ## [0.4.4] - 2018-05-05